Fixed all kQuant and filtered method issues

+++ src/Device/Device.cpp
@@ -11,9 +11,11 @@ void Device::Anderson_disorder(r_type disorder_vec[]){
   size_t SUBDIM = device_vars_.SUBDIM_;
   r_type str = device_vars_.dis_str_;
   
-  for(int i=0;i<SUBDIM; i++){
+  
+  for(size_t i=0;i<SUBDIM; i++){
     r_type random_potential = str * this->rng().get()-str/2;
 
+    
     disorder_vec[i] = random_potential;
   }
   
@@ -27,9 +29,10 @@ void Device::Anderson_disorder(){
   size_t SUBDIM = device_vars_.SUBDIM_;
   r_type str = device_vars_.dis_str_;
   
-  for(int i=0;i<SUBDIM; i++){
+  
+  for(size_t i=0;i<SUBDIM; i++){
     r_type random_potential = str * this->rng().get()-str/2;
-
+    
     dis_[i] = random_potential;
   }
